named storages are appear

+++ evernotebot/wsgi.py
@@ -3,7 +3,7 @@ import hashlib
 import logging
 import traceback
 from time import time
-from os.path import dirname, realpath, join
+from os.path import dirname, realpath, join, exists
 from urllib.parse import urlparse
 
 from uhttp import WsgiApplication
@@ -15,16 +15,18 @@ from evernotebot.bot.core import EvernoteBot
 from evernotebot.bot.shortcuts import evernote_oauth_callback
 
 
-def _save_failed_update(dirpath: str, data: dict):
+def _save_failed_update(dirpath: str, data: dict, str_exc: str):
     current_time = time()
-    key = "{time}{data}".format(time=current_time, data=data).encode()
+    key = f"{current_time}{data}".encode()
     key_hash = hashlib.sha1(key).hexdigest()
-    filename = join(dirpath, "{hash}.json".format(hash=key_hash))
+    filename = join(dirpath, f"{key_hash}.json")
     failed_update = {
         "created": current_time,
         "data": data,
-        "exception": traceback.format_exc(),
+        "exception": str_exc,
     }
+    if exists(filename):
+        raise FileExistsError(filename)
     with open(filename, 'w') as f:
         json.dump(failed_update, f)
 
@@ -34,8 +36,9 @@ def telegram_hook(request):
     bot = request.app.bot
     try:
         bot.process_update(data)
-    except Exception as e:
-        _save_failed_update(bot.config["failed_updates_root"], data)
+    except Exception:
+        str_exc = traceback.format_exc()
+        _save_failed_update(bot.config["failed_updates_root"], data, str_exc)
 
 
 def evernote_oauth(request):
